+++ packages/rum/src/index.ts
@@ -1,11 +1,9 @@
-// Keep the following in sync with packages/rum-recorder/src/index.ts
+// Keep the following in sync with packages/rum-slim/src/index.ts
 export { datadogRum } from './boot/rum.entry'
 export {
   CommonProperties,
-  ProvidedSource,
   RumPublicApi as RumGlobal,
   RumInitConfiguration,
-  RumUserConfiguration,
   // Events
   RumEvent,
   RumActionEvent,