+++ test/e2e/lib/framework/serverApps/mock.ts
@@ -1,7 +1,7 @@
 import * as url from 'url'
 import cors from 'cors'
 import express from 'express'
-import { buildLogs, buildNpm, buildRum, buildRumRecorder, Endpoints } from '../sdkBuilds'
+import { buildLogs, buildNpm, buildRum, buildRumSlim, Endpoints } from '../sdkBuilds'
 
 export function createMockServerApp(endpoints: Endpoints, setup: string) {
   const app = express()
@@ -59,8 +59,8 @@ export function createMockServerApp(endpoints: Endpoints, setup: string) {
     res.header('content-type', 'application/javascript').send(await buildRum(endpoints))
   })
 
-  app.get('/datadog-rum-recorder.js', async (_req, res) => {
-    res.header('content-type', 'application/javascript').send(await buildRumRecorder(endpoints))
+  app.get('/datadog-rum-slim.js', async (_req, res) => {
+    res.header('content-type', 'application/javascript').send(await buildRumSlim(endpoints))
   })
 
   app.get('/app.js', async (_req, res) => {