+++ test/e2e/scenario/recorder.scenario.ts
@@ -1,5 +1,6 @@
-import { CreationReason, IncrementalSource, Segment } from '@datadog/browser-rum-recorder/cjs/types'
-import { InputData, StyleSheetRuleData, NodeType } from '@datadog/browser-rum-recorder/cjs/domain/record/types'
+import { CreationReason, IncrementalSource, Segment } from '@datadog/browser-rum/cjs/types'
+import { InputData, StyleSheetRuleData, NodeType } from '@datadog/browser-rum/cjs/domain/record/types'
+import { RumInitConfiguration } from '@datadog/browser-rum-core'
 
 import { createTest, bundleSetup, html, EventRegistry } from '../lib/framework'
 import { browserExecute } from '../lib/helpers/browser'
@@ -13,7 +14,7 @@ import {
   findMeta,
   findTextContent,
   createMutationPayloadValidatorFromSegment,
-} from '../../../packages/rum-recorder/test/utils'
+} from '../../../packages/rum/test/utils'
 
 const INTEGER_RE = /^\d+$/
 const TIMESTAMP_RE = /^\d{13}$/
@@ -21,7 +22,8 @@ const UUID_RE = /^[0-9a-f]{8}-([0-9a-f]{4}-){3}[0-9a-f]{12}$/
 
 describe('recorder', () => {
   createTest('record mouse move')
-    .withRumRecorder()
+    .withRum()
+    .withRumInit(initRumAndStartRecording)
     .run(async ({ events }) => {
       await browserExecute(() => document.documentElement.outerHTML)
       const html = await $('html')
@@ -66,8 +68,9 @@ describe('recorder', () => {
 
   describe('full snapshot', () => {
     createTest('obfuscate elements')
+      .withRum()
+      .withRumInit(initRumAndStartRecording)
       .withSetup(bundleSetup)
-      .withRumRecorder()
       .withBody(
         html`
           <my.visible.el id="not-obfuscated">foo</my.visible.el>
@@ -115,8 +118,9 @@ describe('recorder', () => {
 
   describe('mutations observer', () => {
     createTest('record mutations')
+      .withRum()
+      .withRumInit(initRumAndStartRecording)
       .withSetup(bundleSetup)
-      .withRumRecorder()
       .withBody(
         html`
           <p>mutation observer</p>
@@ -161,8 +165,9 @@ describe('recorder', () => {
       })
 
     createTest('record character data mutations')
+      .withRum()
+      .withRumInit(initRumAndStartRecording)
       .withSetup(bundleSetup)
-      .withRumRecorder()
       .withBody(
         html`
           <p>mutation observer</p>
@@ -213,8 +218,9 @@ describe('recorder', () => {
       })
 
     createTest('record attributes mutations')
+      .withRum()
+      .withRumInit(initRumAndStartRecording)
       .withSetup(bundleSetup)
-      .withRumRecorder()
       .withBody(
         html`
           <p>mutation observer</p>
@@ -257,8 +263,9 @@ describe('recorder', () => {
       })
 
     createTest("don't record hidden elements mutations")
+      .withRum()
+      .withRumInit(initRumAndStartRecording)
       .withSetup(bundleSetup)
-      .withRumRecorder()
       .withBody(
         html`
           <div data-dd-privacy="hidden">
@@ -284,8 +291,9 @@ describe('recorder', () => {
       })
 
     createTest('record DOM node movement 1')
+      .withRum()
+      .withRumInit(initRumAndStartRecording)
       .withSetup(bundleSetup)
-      .withRumRecorder()
       .withBody(
         // prettier-ignore
         html`
@@ -332,8 +340,9 @@ describe('recorder', () => {
       })
 
     createTest('record DOM node movement 2')
+      .withRum()
+      .withRumInit(initRumAndStartRecording)
       .withSetup(bundleSetup)
-      .withRumRecorder()
       .withBody(
         // prettier-ignore
         html`
@@ -383,8 +392,9 @@ describe('recorder', () => {
       })
 
     createTest('serialize node before record')
+      .withRum()
+      .withRumInit(initRumAndStartRecording)
       .withSetup(bundleSetup)
-      .withRumRecorder()
       .withBody(
         // prettier-ignore
         html`
@@ -436,8 +446,9 @@ describe('recorder', () => {
 
   describe('input observers', () => {
     createTest('record input interactions')
+      .withRum()
+      .withRumInit(initRumAndStartRecording)
       .withSetup(bundleSetup)
-      .withRumRecorder()
       .withBody(
         html`
           <form>
@@ -515,8 +526,9 @@ describe('recorder', () => {
       })
 
     createTest("don't record ignored input interactions")
+      .withRum()
+      .withRumInit(initRumAndStartRecording)
       .withSetup(bundleSetup)
-      .withRumRecorder()
       .withBody(
         html`
           <input type="text" id="first" name="first" />
@@ -551,8 +563,9 @@ describe('recorder', () => {
       })
 
     createTest('replace masked values by asterisks')
+      .withRum()
+      .withRumInit(initRumAndStartRecording)
       .withSetup(bundleSetup)
-      .withRumRecorder()
       .withBody(
         html`
           <input type="text" id="by-data-attribute" data-dd-privacy="input-masked" />
@@ -584,8 +597,9 @@ describe('recorder', () => {
 
   describe('stylesheet rules observer', () => {
     createTest('record dynamic CSS changes')
+      .withRum()
+      .withRumInit(initRumAndStartRecording)
       .withSetup(bundleSetup)
-      .withRumRecorder()
       .withBody(
         html`
           <style>
@@ -621,4 +635,9 @@ describe('recorder', () => {
 
 function getFirstSegment(events: EventRegistry) {
   return events.sessionReplay[0].segment.data
+}
+
+function initRumAndStartRecording(initConfiguration: RumInitConfiguration) {
+  window.DD_RUM!.init(initConfiguration)
+  window.DD_RUM!.startSessionReplayRecording()
 }