+++ CHANGELOG.md
@@ -16,6 +16,68 @@
 
 ---
 
+## v3.0.0
+
+Here is the list of all the breaking changes between version 2 and 3 of the RUM SDK.
+The Log SDK isn't impacted and is still fully backward compatible.
+
+For each change, we show the new way to handle your use case.
+
+### Network errors no longer monitored
+
+RUM browser SDK no longer automatically monitors network errors.
+
+**New option**: You can use the option beforeSend to get the resource, check the `status` and send it with the `addError()` API.
+
+### Source attribute of addError() has been removed
+
+You can no longer change the source of error sent with `addError()`. All these errors have a `custom` source.
+
+**New option**: If you used this feature to identify `network` and `source` errors, you can add context attributes with `addError()` instead.
+
+### New intake domains
+
+RUM now uses a new intake domain. Therefore the init options `useAlternateIntakeDomains` has been removed.
+
+| Old domains                        | New domains                      |
+| ---------------------------------- | -------------------------------- |
+| rum-http-intake.logs.datadoghq.com | rum.browser-intake-datadoghq.com |
+| rum-http-intake.logs.datadoghq.eu  | rum.browser-intake-datadoghq.eu  |
+
+> Warning: You have to update your content security policy, if you use one.
+
+### Removed typescript types
+
+| Old types                    | New types                    |
+| ---------------------------- | ---------------------------- |
+| RumUserConfiguration         | RumInitConfiguration         |
+| RumRecorderUserConfiguration | RumRecorderInitConfiguration |
+
+### Removed init options
+
+| Old options        | New options |
+| ------------------ | ----------- |
+| publicApiKey       | clientToken |
+| datacenter         | site        |
+| resourceSampleRate | NONE        |
+
+### Removed API
+
+| Old API       | New API   |
+| ------------- | --------- |
+| addUserAction | addAction |
+
+### Changes
+
+- 💥 always use alternative domains for RUM ([#944](https://github.com/DataDog/browser-sdk/pull/944))
+- 💥 remove deprecated rum init options ([#940](https://github.com/DataDog/browser-sdk/pull/940))
+- 💥 remove deprecated types ([#942](https://github.com/DataDog/browser-sdk/pull/942))
+- 💥 [RUMF-951] merge rum-recorder into rum ([#941](https://github.com/DataDog/browser-sdk/pull/941))
+- 👷 [RUMF-937] create the rum-slim package ([#935](https://github.com/DataDog/browser-sdk/pull/935))
+- 💥 remove deprecated RUM.addUserAction ([#939](https://github.com/DataDog/browser-sdk/pull/939))
+- ✨ [RUMF-940] implement the replay sample rate option and remove resource sample rate option ([#931](https://github.com/DataDog/browser-sdk/pull/931))
+- 💥 Remove addError 'source' argument ([#936](https://github.com/DataDog/browser-sdk/pull/936))
+
 ## v2.18.0
 
 - 👷 Versioned CDN files ([#932](https://github.com/DataDog/browser-sdk/pull/932))